requirements.txt
async-generator==1.10
attrs==22.1.0
black==22.6.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
et-xmlfile==1.1.0
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
google-api-core==2.8.2
google-api-python-client==2.56.0
google-auth==2.10.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
h11==0.13.0
httplib2==0.20.4
huggingface-hub==0.8.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
numpy==1.23.1
oauthlib==3.2.0
openai==0.22.1
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.4.3
pandas-stubs==1.4.3.220807
pathspec==0.9.0
platformdirs==2.5.2
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.2
PyYAML==6.0
regex==2022.7.25
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tokenizers==0.12.1
tomli==2.0.1
torch==1.12.1
tqdm==4.64.0
transformers==4.21.1
trio==0.21.0
trio-websocket==0.9.2
types-pytz==2022.1.2
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
Werkzeug==2.2.2
wsproto==1.2.0
yapf==0.32.0
zipp==3.8.1